+++ Sources/Sentry/Profiling/SentryContinuousProfiler.mm
@@ -5,7 +5,6 @@
 #    import "SentryDependencyContainer.h"
 #    import "SentryLogC.h"
 #    import "SentryMetricProfiler.h"
-#    import "SentryNSNotificationCenterWrapper.h"
 #    import "SentryNSTimerFactory.h"
 #    import "SentryProfiler+Private.h"
 #    import "SentryProfilerSerialization.h"
@@ -132,7 +131,9 @@ _sentry_unsafe_stopTimerAndCleanup()
                      queue:nil
                 usingBlock:^(NSNotification *_Nonnull notification) {
                     [SentryDependencyContainer.sharedInstance.notificationCenterWrapper
-                        removeObserver:_observerToken];
+                        removeObserver:_observerToken
+                                  name:nil
+                                object:nil];
                     [self stopTimerAndCleanup];
                 }];
 #    endif // SENTRY_HAS_UIKIT
@@ -225,7 +226,9 @@ _sentry_unsafe_stopTimerAndCleanup()
 #    if SENTRY_HAS_UIKIT
     if (_observerToken != nil) {
         [SentryDependencyContainer.sharedInstance.notificationCenterWrapper
-            removeObserver:_observerToken];
+            removeObserver:_observerToken
+                      name:nil
+                    object:nil];
     }
 #    endif // SENTRY_HAS_UIKIT
 
@@ -260,7 +263,9 @@ _sentry_unsafe_stopTimerAndCleanup()
 #        if SENTRY_HAS_UIKIT
     if (_observerToken != nil) {
         [SentryDependencyContainer.sharedInstance.notificationCenterWrapper
-            removeObserver:_observerToken];
+            removeObserver:_observerToken
+                      name:UIApplicationWillResignActiveNotification
+                    object:nil];
         _observerToken = nil;
     }
 #        endif // SENTRY_HAS_UIKIT