+++ Sources/Sentry/Public/SentrySDK.h
@@ -253,6 +253,7 @@ SENTRY_NO_INIT
               withScopeBlock:(void (^)(SentryScope *scope))block
     NS_SWIFT_NAME(capture(message:block:));
 
+#if !SDK_V9
 /**
  * Captures user feedback that was manually gathered and sends it to Sentry.
  * @param userFeedback The user feedback to send to Sentry.
@@ -263,6 +264,7 @@ SENTRY_NO_INIT
     NS_SWIFT_NAME(capture(userFeedback:)) DEPRECATED_MSG_ATTRIBUTE(
         "Use SentrySDK.captureFeedback or use or configure our new managed UX with "
         "SentryOptions.configureUserFeedback.");
+#endif // !SDK_V9
 
 /**
  * Captures user feedback that was manually gathered and sends it to Sentry.