+++ Sources/Swift/Helper/SentryEnabledFeaturesBuilder.swift
@@ -3,6 +3,7 @@ import Foundation
 @objcMembers @_spi(Private) public class SentryEnabledFeaturesBuilder: NSObject {
 
     // swiftlint:disable cyclomatic_complexity function_body_length
+    @available(*, deprecated, message: "This is only marked as deprecated because enableAppLaunchProfiling is marked as deprecated. Once that is removed this can be removed.")
     public static func getEnabledFeatures(options: Options?) -> [String] {
         guard let options = options else {
             return []