+++ Tests/SentryProfilerTests/SentryAppStartProfilingConfigurationChangeTests.swift
@@ -5,6 +5,7 @@ import XCTest
 
 #if os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
 /// Validate stopping behavior of launch profiles that run with one set of configured options, where the SDK is started on that launch with a different set of options, to validate that the configured options persisted to disk from the previous launch are the ones used to determine how/when to stop the profiler, and not the new options currently in memory
+@available(*, deprecated, message: "This is only marked as deprecated because it must still test some deprecated profiling APIs, but the deprecation warnings are converted to errors in our test targets. Once the deprecated API are removed, this can also be removed.")
 final class SentryAppStartProfilingConfigurationChangeTests: XCTestCase {
     private var fixture: SentryProfileTestFixture!
 
@@ -20,6 +21,7 @@ final class SentryAppStartProfilingConfigurationChangeTests: XCTestCase {
 }
 
 // MARK: configuration changes between launches (no TTFD combinations, see iOS-only tests)
+@available(*, deprecated, message: "This is only marked as deprecated because it must still test some deprecated profiling APIs, but the deprecation warnings are converted to errors in our test targets. Once the deprecated API are removed, this can also be removed.")
 extension SentryAppStartProfilingConfigurationChangeTests {
     func test_lastLaunch_traceBased_currentLaunch_continuousV1() throws {
         // Arrange
@@ -384,6 +386,7 @@ extension SentryAppStartProfilingConfigurationChangeTests {
 
 #if !os(macOS)
 // MARK: configuring launch profiling with TTFD disabled, then launching with it enabled (iOS-only)
+@available(*, deprecated, message: "This is only marked as deprecated because it must still test some deprecated profiling APIs, but the deprecation warnings are converted to errors in our test targets. Once the deprecated API are removed, this can also be removed.")
 extension SentryAppStartProfilingConfigurationChangeTests {
     // MARK: starting with trace-based no TTFD
     func test_lastLaunch_traceBased_noTTFD_currentLaunch_continuousV1_withTTFD() throws {
@@ -933,6 +936,7 @@ extension SentryAppStartProfilingConfigurationChangeTests {
 }
 
 // MARK: configuring launch profiling with TTFD enabled, then launching with it disabled (iOS-only)
+@available(*, deprecated, message: "This is only marked as deprecated because it must still test some deprecated profiling APIs, but the deprecation warnings are converted to errors in our test targets. Once the deprecated API are removed, this can also be removed.")
 extension SentryAppStartProfilingConfigurationChangeTests {
     // MARK: starting with trace-based with TTFD
     func test_lastLaunch_traceBased_withTTFD_currentLaunch_continuousV1_noTTFD() throws {