+++ Tests/SentryProfilerTests/SentryProfilingPublicAPITests.swift
@@ -68,6 +68,7 @@ class SentryProfilingPublicAPITests: XCTestCase {
 }
 
 // MARK: transaction profiling
+@available(*, deprecated, message: "Transaction profiling is deprecated")
 extension SentryProfilingPublicAPITests {
     func testSentryOptionsReportsProfilingCorrelatedToTraces_NonnilSampleRate() {
         // Arrange
@@ -103,6 +104,7 @@ extension SentryProfilingPublicAPITests {
 }
 
 // MARK: continuous profiling v1
+@available(*, deprecated, message: "Continuous profiling v1 is deprecated")
 extension SentryProfilingPublicAPITests {
     func testSentryOptionsReportsContinuousProfilingEnabled() {
         // Arrange
@@ -208,6 +210,7 @@ extension SentryProfilingPublicAPITests {
 }
 
 // MARK: continuous profiling v2
+@available(*, deprecated, message: "This is only deprecated because profilesSampleRate is deprecated. Once that is removed this attribute can be removed.")
 extension SentryProfilingPublicAPITests {
     func testSentryOptionsReportsContinuousProfilingV2Enabled() {
         // Arrange